feat: add router for single page application
split users and Pod to seperate routes and components.

add Navbar with links of routes.
